EWPP-297: Split the admin tests for the two entities.
